Switch cli to scala 2.12 and latest case-app
