+++ cli/src/main/scala-2.12/coursier/cli/CoursierCommand.scala
@@ -1,0 +1,26 @@
+package coursier.cli
+
+import caseapp.CommandParser
+import caseapp.core.help.CommandsHelp
+
+object CoursierCommand {
+
+  val parser =
+    CommandParser.nil
+      .add(Bootstrap)
+      .add(Fetch)
+      .add(Launch)
+      .add(Resolve)
+      .add(SparkSubmit)
+      .reverse
+
+  val help =
+    CommandsHelp.nil
+      .add(Bootstrap)
+      .add(Fetch)
+      .add(Launch)
+      .add(Resolve)
+      .add(SparkSubmit)
+      .reverse
+
+}
