+++ cli/src/main/scala-2.12/coursier/cli/Resolve.scala
@@ -1,0 +1,13 @@
+package coursier
+package cli
+
+import caseapp._
+import coursier.cli.options.ResolveOptions
+
+object Resolve extends CaseApp[ResolveOptions] {
+
+  def run(options: ResolveOptions, args: RemainingArgs): Unit = {
+    new Helper(options.common, args.all, printResultStdout = true)
+  }
+
+}
