+++ cli/src/main/scala-2.12/coursier/cli/SoftExcludeParsingException.scala
@@ -1,0 +1,6 @@
+package coursier.cli
+
+final class SoftExcludeParsingException(
+  private val message: String = "",
+  private val cause: Throwable = None.orNull
+) extends Exception(message, cause)
