+++ cli/src/main/scala-2.12/coursier/cli/Util.scala
@@ -1,0 +1,23 @@
+package coursier.cli
+
+object Util {
+
+  def prematureExit(msg: String): Nothing = {
+    Console.err.println(msg)
+    sys.exit(255)
+  }
+
+  def prematureExitIf(cond: Boolean)(msg: => String): Unit =
+    if (cond)
+      prematureExit(msg)
+
+  def exit(msg: String): Nothing = {
+    Console.err.println(msg)
+    sys.exit(1)
+  }
+
+  def exitIf(cond: Boolean)(msg: => String): Unit =
+    if (cond)
+      exit(msg)
+
+}
