+++ cli/src/main/scala-2.12/coursier/cli/options/BootstrapOptions.scala
@@ -1,0 +1,15 @@
+package coursier.cli.options
+
+import caseapp.{Parser, Recurse}
+
+final case class BootstrapOptions(
+  @Recurse
+    artifactOptions: ArtifactOptions,
+  @Recurse
+    options: BootstrapSpecificOptions
+)
+
+object BootstrapOptions {
+  implicit val parser = Parser[BootstrapOptions]
+  implicit val help = caseapp.core.help.Help[BootstrapOptions]
+}
