+++ cli/src/main/scala-2.12/coursier/cli/options/FetchOptions.scala
@@ -1,0 +1,24 @@
+package coursier.cli.options
+
+import caseapp.{ HelpMessage => Help, ExtraName => Short, _ }
+
+final case class FetchOptions(
+  @Help("Fetch source artifacts")
+  @Short("S")
+    sources: Boolean = false,
+  @Help("Fetch javadoc artifacts")
+  @Short("D")
+    javadoc: Boolean = false,
+  @Help("Print java -cp compatible output")
+  @Short("p")
+    classpath: Boolean = false,
+  @Recurse
+    artifactOptions: ArtifactOptions = ArtifactOptions(),
+  @Recurse
+    common: CommonOptions = CommonOptions()
+)
+
+object FetchOptions {
+  implicit val parser = Parser[FetchOptions]
+  implicit val help = caseapp.core.help.Help[FetchOptions]
+}
