+++ cli/src/main/scala-2.12/coursier/cli/options/ResolveOptions.scala
@@ -1,0 +1,8 @@
+package coursier.cli.options
+
+import caseapp._
+
+final case class ResolveOptions(
+  @Recurse
+    common: CommonOptions = CommonOptions()
+)
