+++ cli/src/main/scala-2.12/coursier/cli/scaladex/Scaladex.scala
