+++ cli/src/main/scala-2.12/coursier/cli/util/JsonReport.scala
