+++ cli/src/test/scala-2.12/coursier/cli/BUILD
@@ -2,7 +2,7 @@ junit_tests(
     name = "test",
     dependencies = [
         "3rdparty/jvm:scalatest",
-        "cli/src/main/scala-2.11:cli",
+        "cli/src/main/scala-2.12:cli",
     ],
     sources = globs("*.scala"),
 )
