+++ cli/src/test/scala-2.12/coursier/cli/CliUnitTest.scala
@@ -2,6 +2,7 @@ package coursier.cli
 
 import java.io.{File, FileWriter}
 
+import coursier.cli.options.CommonOptions
 import org.junit.runner.RunWith
 import org.scalatest.FlatSpec
 import org.scalatest.junit.JUnitRunner
