+++ extra/src/main/scala-2.12/coursier/extra/BUILD
