+++ extra/src/main/scala/coursier/extra/BUILD
@@ -2,6 +2,7 @@ scala_library(
     name = "extra",
     dependencies = [
         "core:core",
+        "3rdparty/jvm:scala-native",
     ],
     sources = rglobs("*.scala"),
 )
