+++ project/SharedVersions.scala
@@ -4,7 +4,7 @@ object SharedVersions {
   def asm = "5.2"
   def fastParse = "1.0.0"
   def proguard = "5.3.3"
-  def scalaNative = "0.3.0-coursier-1"
+  def scalaNative = "0.3.6"
   def scalaz = "7.2.16"
 
 }
